Make Shutter work
I had the problem that the shutter was only opening and closing when using the command line with the commands "SHT:0"(close) or "SHT:1"(open) found in the current manual. This also changed the button label. -> So I changed the command accordingly
Also compiled a new executable with MSYS2 which works on my Win11 computer but not sure if it works everywhere (I am absolutely new to C++ and Makefile-compilation)